VS 2013 ADD_MSVC_PRECOMPILED_HEADER compat fix
Fixed ADD_MSVC_PRECOMPILED_HEADER so it doesn't break the config under VS 2013.

This breakage was due to the fact that the source file added wasn't pathed, which resulted in a .obj being added as a dependency and it not adding the precompiled header create flag to the specified file.

The method now deals with this as well as not requiring the caller to remove said file from sources.